# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthProvider": {
        "enum": [
          "magic",
          "boostify",
          "email",
          "google",
          "apple",
          "guest"
        ],
        "title": "AuthProvider",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProfileUpdate": {
        "properties": {
          "avatar": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "ProfileUpdate",
        "type": "object"
      },
...
      "PublicUser": {
        "description": "Shape returned to the frontend (no secrets).",
        "properties": {
          "avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "provider": {
            "$ref": "#/components/schemas/AuthProvider"
          }
        },
        "required": [
          "id",
          "provider",
          "created_at"
        ],
        "title": "PublicUser",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Identity, wallet and AI content generation for TOONTOON.",
    "title": "ARTEKI API",
    "version": "generated"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/profile": {
      "delete": {
        "description": "Delete the account.\n\nThis is **anonymisation, not a DELETE**: the ledger references the user by\nforeign key and the database rightly refuses to erase a row that money\nhistory points at. So the person disappears — email, name and avatar are\ncleared, the session is killed — while the bookkeeping stays intact (CH-15).\n\nФайлы стираются здесь же. До 31 августа 2026 не стирались: строка\nобезличивалась, а снимки оставались лежать в хранилище. Человек нажимал\n«удалить», ему отвечали «готово», и его лицо продолжало храниться у нас.",
        "operationId": "delete_profile_api_auth_profile_delete",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "arteki-session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Arteki-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Profile",
        "tags": [
          "profile"
        ]
      },
      "patch": {
        "operationId": "update_profile_api_auth_profile_patch",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "arteki-session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Arteki-Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUser"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile",
        "tags": [
          "profile"
        ]
      }
    },
...
  }
}